fix CI in newer rust
The culprit lied in rust-lang/rust#116505

In short, with default features turned off, main was trivial enough to
be marked as inline function automatically which then made the symbol
weak. Since nobody was referencing it, it got stripped away.

Marking main in default-features=false config as #[inline(never)] or
replacing 1 + 1 in it's body with a simple println!("foo") call (to make
the main function sophisticated enough not to be subject to the new
automatic marking as inline) makes the test pass again.