Add a `-s`/`--silent` option to silence user-facing messages
I have a usecase where it would be nice to use cargo-show-asm
programmatically, since it does a nicer job of filtering and demangling
than most builtin utilities. However, UI messages currently print to
stdout, meaning they need to be manually filtered:

    $ cargo asm -p libm --lib 2>/dev/null | cat
    Try one of those by name or a sequence number
      0 "<f32 as libm::math::support::float_traits::Float>::abs" [9]
      1 "<f32 as libm::math::support::float_traits::Float>::copysign" [18]
      2 "<f32 as libm::math::support::float_traits::Float>::normalize" [21]
      3 "<f64 as libm::math::support::float_traits::Float>::abs" [9]

Introduce a `-s`/`--silent` flag that reduces the verbosity level,
effectively being the the opposite of `-v`:

    $ path/cargo-asm asm -p libm --lib -s 2>/dev/null | cat
      0 "<f32 as libm::math::support::float_traits::Float>::abs" [9]
      1 "<f32 as libm::math::support::float_traits::Float>::copysign" [18]
      2 "<f32 as libm::math::support::float_traits::Float>::normalize" [21]
      3 "<f64 as libm::math::support::float_traits::Float>::abs" [9]

`-s` was chosen to mirror Curl, since `-q` is already used as the flag
to quieten Cargo's output.